+++ client/web/src/enterprise/insights/components/insights-view-grid/components/backend-insight/components/drill-down-filters-panel/drill-down-filters/DrillDownInsightFilters.tsx
@@ -161,8 +161,9 @@ export const DrillDownInsightFilters: FunctionComponent<DrillDownInsightFilters>
                     variant="link"
                     className={classNames(styles.actionButton, styles.actionButtonWithCollapsed)}
                     onClick={() => onVisualModeChange(FilterSectionVisualMode.HorizontalSections)}
+                    aria-label="Switch to horizontal mode"
                 >
-                    <Icon as={ArrowExpandIcon} />
+                    <Icon as={ArrowExpandIcon} aria-hidden={true} />
                 </Button>
             </header>
         )
@@ -189,14 +190,15 @@ export const DrillDownInsightFilters: FunctionComponent<DrillDownInsightFilters>
                         variant="link"
                         className={styles.actionButton}
                         onClick={() => onVisualModeChange(FilterSectionVisualMode.Preview)}
+                        aria-label="Switch to preview mode"
                     >
-                        <Icon as={ArrowCollapseIcon} />
+                        <Icon as={ArrowCollapseIcon} aria-hidden={true} />
                     </Button>
                 )}
             </header>
             <hr className={styles.headerSeparator} />
 
-            <div className={classNames(styles.panels, { [styles.panelsHorizontalMode]: isHorizontalMode })}>
+            <div className={classNames({ [styles.panelsHorizontalMode]: isHorizontalMode })}>
                 {showSeriesDisplayOptions && (
                     <FilterCollapseSection
                         open={isHorizontalMode || activeSection === FilterSection.SortFilter}
@@ -205,6 +207,7 @@ export const DrillDownInsightFilters: FunctionComponent<DrillDownInsightFilters>
                         preview={getSortPreview(parseSeriesDisplayOptions(seriesDisplayOptions))}
                         hasActiveFilter={hasSeriesDisplayOptionsChanged}
                         withSeparators={!isHorizontalMode}
+                        className={classNames(styles.panel, { [styles.panelHorizontalMode]: isHorizontalMode })}
                         onOpenChange={opened => handleCollapseState(FilterSection.SortFilter, opened)}
                     >
                         <SortFilterSeriesPanel